Fix:MainPage 분리 컴포넌트 일부 분리
