+++ src/hooks/useVisiblity.ts
@@ -1,0 +1,23 @@
+import { useState, useEffect } from "react";
+
+const useVisibility = (elementId: string) => {
+  const [isVisible, setIsVisible] = useState(false);
+
+  useEffect(() => {
+    const handleScroll = () => {
+      const element = document.getElementById(elementId);
+      if (element && element.getBoundingClientRect().top < window.innerHeight) {
+        setIsVisible(true);
+      }
+    };
+
+    window.addEventListener("scroll", handleScroll);
+    handleScroll();
+
+    return () => window.removeEventListener("scroll", handleScroll);
+  }, [elementId]);
+
+  return isVisible;
+};
+
+export default useVisibility;
